Update ServerNode.java
This is a major update for the ServerNode object which can now handle a wide array of different types of client connections. So far it has been tested with the Java Client, PHP WebSocket, and ESP8266WiFi / ESP8266WiFiMulti in Arduino on a Feather HUZZAH ESP8266 Breakout board. I wanted to specifically test these three first because it opens the doors for connections with cross-platform software, website applications, and physical devices that have internet capabilities.